refactor(core): separate core into {fixer, linter}-processor
It aim to remove core logic from core
